|UPDATE| Theme switcher now works
