--- common/components/navigationbar.bloc.ts
@@ -1,15 +1,0 @@
-import { useLayoutContext } from '../../context/LayoutContext';
-import { useErrorHandler } from '../hooks/error-handler';
-
-const useNavigationBar = () => {
-  //hooks
-  const {} = useErrorHandler();
-  const { theme, setTheme } = useLayoutContext();
-
-  //handlers
-  function onThemeSwitchHandler() {
-    setTheme(theme == 'dark' ? 'light' : 'dark');
-  }
-  return { onThemeSwitchHandler, theme };
-};
-export { useNavigationBar };
