+++ common/enums/index.ts
@@ -1,0 +1,3 @@
+export enum StorageKey {
+  THEME = 'theme',
+}
